Refactoring Programe.cs And Add Extensions Folders
