+++ MarketPOS.API/MarketPOS.API/Extensions/ExtensionValidatuion/ValidationExtensions.cs
@@ -1,0 +1,11 @@
+﻿namespace MarketPOS.API.Extensions.ExtensionValidatuion;
+
+public static class ValidationExtensions
+{
+    public static IServiceCollection AddCustomValidation(this IServiceCollection services)
+    {
+        services.AddFluentValidationAutoValidation();
+        services.AddValidatorsFromAssemblies(AppDomain.CurrentDomain.GetAssemblies());
+        return services;
+    }
+}
