+++ MarketPOS.API/MarketPOS.API/ExtensionsFiltreingAndMiddlewares/Exceptions/ExtendedProblemDetails.cs
