+++ MarketPOS.API/MarketPOS.API/ExtensionsFiltreingAndMiddlewares/FuatuersFunction/HelperMethod.cs
