+++ MarketPOS.API/MarketPOS.Infrastructure/ImplmentationCacheing/Decorator/GenericServiceCacheing.cs
@@ -1,5 +1,5 @@
 ﻿using MarketPOS.Application.InterfaceCacheing;
-using MarketPOS.Infrastructure;
+using MarketPOS.Infrastructure.TrackingServicesMiddleware;
 using Microsoft.EntityFrameworkCore.Query;
 using System.Security.Cryptography;
 using System.Text;
