+++ MarketPOS.API/MarketPOS.Infrastructure/TrackingServicesMiddleware/ServiceTrackingMiddlewareExtensions.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Builder;
+
+namespace MarketPOS.Infrastructure.TrackingServicesMiddleware;
+
+public static class ServiceTrackingMiddlewareExtensions
+{
+    public static IApplicationBuilder UseServiceTracking(this IApplicationBuilder builder)
+    {
+        return builder.UseMiddleware<ServiceTrackingMiddleware>();
+    }
+}
+
+
